refactor: Re-position notice using filter
