+++ src/FormBuilder/resources/js/form-builder/src/blocks/fields/payment-gateways/Edit.tsx
@@ -28,8 +28,6 @@ const GatewayItem = ({label, icon}: {label: string; icon: ReactNode}) => {
 export default function Edit(props: BlockEditProps<any>) {
     const {gateways} = getFormBuilderWindowData();
 
-    const [showNotification, onDismissNotification] = useAdditionalPaymentGatewaysNotice()
-
     return (
         <div
             style={{
@@ -84,17 +82,6 @@ export default function Edit(props: BlockEditProps<any>) {
                         {__('Enable more payment gateways', 'give')}
                     </a>
                 </div>
-                {showNotification && (
-                    <PanelRow>
-                        <InspectorNotice
-                            title={__('Additional Payment Gateways', 'give')}
-                            description={__('Enable multiple payment gateways on your forms via the global settings.', 'give')}
-                            helpText={__('Go to payment gateway settings', 'give')}
-                            helpUrl={'/wp-admin/edit.php?post_type=give_forms&page=give-settings&tab=gateways&section=gateways-settings&group=v3'}
-                            onDismiss={onDismissNotification}
-                        />
-                    </PanelRow>
-                )}
             </InspectorControls>
         </div>
     );
