--- src/FormBuilder/resources/js/form-builder/src/blocks/fields/payment-gateways/hooks/useAdditionalPaymentGatewaysNotice.ts
@@ -1,22 +1,0 @@
-import {useState} from "react";
-
-declare const window: {
-    additionalPaymentGatewaysNotificationData: {
-        actionUrl: string;
-        isDismissed: boolean;
-    };
-} & Window;
-
-/**
- * @unreleased
- */
-const useAdditionalPaymentGatewaysNotice = () => {
-
-    const {actionUrl, isDismissed} = window.additionalPaymentGatewaysNotificationData;
-    const [showNotification, setShowNotification] = useState(!isDismissed);
-    const onDismissNotification = () => fetch(actionUrl, {method: 'POST'}).then(() => setShowNotification(false))
-
-    return [showNotification, onDismissNotification];
-}
-
-export default useAdditionalPaymentGatewaysNotice;
