+++ src/FormBuilder/resources/js/form-builder/src/blocks/fields/payment-gateways/index.tsx
@@ -1,9 +1,13 @@
 import settings from './settings';
 import {FieldBlock} from '@givewp/form-builder/types';
+import {addFilter} from "@wordpress/hooks";
+import withAdditionalPaymentGatewayNotice from './withAdditionalPaymentGatewayNotice';
 
 const paymentGateways: FieldBlock = {
     name: 'givewp/payment-gateways',
     settings,
 };
+
+addFilter('editor.BlockEdit', 'givewp/stripe-payment-element', withAdditionalPaymentGatewayNotice, 999);
 
 export default paymentGateways;
